refactor: move error classes to dedicated types folder and update imports (#18.0.0)
- Relocated `ApiError`, `BaseApiError`, `QuizApiError`, and `TokenUrlError` to `src/types/errors-types.ts`.
- Updated error constant declarations to `src/constants/errors.constants.ts`.
- Adjusted imports across the app to reflect the new structure.
- Reformatted function calls in `startTrivia` and `fetchQuestions` for readability.
- Fixed version typo in `package.json` and `package-lock.json` (from `18s.0.0` to `18.0.0`).

No functional changes; internal code structure improved for better maintainability.
