+++ src/_lib/getFetchQuestionsUrl.ts
@@ -1,8 +1,9 @@
 import DOMPurify from 'dompurify';
 
 import { API_URL, CATEGORY_ID } from '../constants/api.constants';
+import { API_ERRORS } from '../constants/errors.constants.ts';
 import { Difficulty } from '../types/difficulty.enum';
-import { API_ERRORS, QuizApiError } from './errors';
+import { QuizApiError } from '../types/errors-types.ts';
 import { createApiUrlWithErrorHandling } from './url';
 
 interface ValidationConfig {
